test : 각 테스트 분리 및 위치 변경
